Update Merino to grab the talk to the Elastico Elasticsearch
Elastico requires a Cloud ID and API key to be passed to the client.
